feat: container에서 로직을 관리하도록 수정
